meta-aspeed: linux-aspeed: fix SRC_URI to append
Modifications to the SRC_URI should be append rather than a set
operation, so that they do not erase any changed to the SRC_URI that may
have happened prior to the include file.  Otherwise, it can ignore local
patches being applied for testing, which can make for irate developers.

Signed-off-by: Patrick Williams <patrick@stwcx.xyz>
Change-Id: I9bd95352fa30bd0e8dfd89404d2ac0dbb4d78a3a