meta-openpower: remove fsi dependency with occ service
Adding this dependency caused an ordering cycle in certain BMC reboot
paths:

Feb 16 06:24:00 p10bmc systemd[1]: multi-user.target: Found dependency on multi-user.target/start
Feb 16 06:24:00 p10bmc systemd[1]: multi-user.target: Job org.open_power.OCC.Control.service/start deleted to break ordering cycle starting with multi-user.target/start
Feb 16 06:24:00 p10bmc systemd[1]: multi-user.target: Found ordering cycle on org.open_power.OCC.Control.service/start
Feb 16 06:24:00 p10bmc systemd[1]: multi-user.target: Found dependency on fsi-scan@0.service/start
Feb 16 06:24:00 p10bmc systemd[1]: multi-user.target: Found dependency on obmc-power-on@0.target/start
Feb 16 06:24:00 p10bmc systemd[1]: multi-user.target: Found dependency on multi-user.target/start
Feb 16 06:24:00 p10bmc systemd[1]: multi-user.target: Job fsi-scan@0.service/start deleted to break ordering cycle starting with multi-user.target/start

A better overall solution to what was trying to be solved with this
dependency can be found in the following commit:
https://github.com/openbmc/openpower-occ-control/commit/1718fd8bcd18f93accb6ed87b36f4c768a93f61a

The problem we're seeing is that because we have to run the
fsi-scan.service in scenarios where we reboot the BMC while the host
is up, the occ application sees the OCC's for a period of time
before the fsi-scan runs, but then they "disappear" for a bit while
the fsi-scan is run. It's complicated for the app to handle this.
We've tried a variety of solutions to try and get the occ
application to handle this, but in the end we decided that since the
occ app is not required to be running immediately, we'll just have
it wait for the r/r process (and the fsi-scan service) to complete
via that other commit.

Tested:
- Set APR policy to always power on and rebooted the BMC. Verified I did
  not see any "Found dependency" errors and system booted fine.

Signed-off-by: Andrew Geissler <geissonator@yahoo.com>
Change-Id: Ide33a9ec21e8eaddc8734351600931d961f811ca