Fix the bug that devtool modify u-boot is not picked up by image-bmc
Previously when u-boot code was changed inside a devtool modify repo, the
u-boot artifacts were updated but they could not be picked up by the
final image-bmc. The reason was that the sstate version of u-boot was
used because the dependency between u-boot and obmc-phosphor-image was
not set up properly.

Fix the issue by adding the proper inter-task dependency between the
two.

Tested:
The image-bmc md5sum changes as expected when u-boot code is changed.

Change-Id: Ic98d3097f66e351c774ced4ee435149b5441cf67
Signed-off-by: Yong Zhao <yozhao@google.com>