meta-phosphor: Add option to enable cgroup support
We've noticed some poor behavior of systemd processes
on the Meta BMCs. We enabled these kernel configurations
to ensure we get memory limits adhered to and can easily
debug things.

Test Plan: Verified that this configuration was picked
up in YV4 and that when this is not specified, none of
these new kernel config options are picked up.

Change-Id: I2d60680fb7df52ce50cf88741b0e35809690da31
Signed-off-by: James Athappilly <jamesatha@gmail.com>