meta-ampere: ampere-ipmi-oem: add RDEPENDS on openssl binary
We are using openssl command to create the certificate fingerprint [1],
but there is no dependency listed in the recipe.

This adds RDEPENDS on the openssl binary.

Long term solution should be linking to the openssl library directly,
but as we intend to replace this current implementation with something
else different [2], this change should suffice in the meanwhile.

Tested:

- openssl command executes successfully:

    root@mtmitchell-dcscm:~# openssl -v
    OpenSSL 3.4.0 22 Oct 2024 (Library: OpenSSL 3.4.0 22 Oct 2024)

- Get Manager Certificate Fingerprint via IPMI inband successfully:

    [root@fedora ~]# ipmitool raw 0x2c 0x01 0x52 0x01
     52 01 d4 58 4d 68 1c 78 8f 94 3d 48 98 a1 fa 62
     38 bf d1 7e 85 9a bf 7a 3f dc 80 a3 16 eb ed 6d
     d5 f5

[1] https://github.com/ampere-openbmc/ampere-ipmi-oem/blob/f92b584da2253797d17ba8c4af8c6f1cbe28dad2/src/redfishhostinterface.cpp#L264-L271
[2] https://gerrit.openbmc.org/c/openbmc/docs/+/79327

Change-Id: I1292ddbe96ac7ca08a85d81686115fde83516723
Signed-off-by: Khang D Nguyen <khangng@os.amperecomputing.com>