meta-google: gbmc-bridge: Add version to DHCP req
We want to know what version of the BMC is running as soon as possible
when the netboot process starts. DHCP messages are the first things we
see when a BMC is brought up and we have unused request fields that can
be use for the purpose of conveying the BMC version.

Tested: Verified via an install job that the field could be represented
nicely in the log.

Change-Id: Ibd56999acf483060f674db46cabb2a465392fb97
Signed-off-by: William A. Kennington III <wak@google.com>