pldm: split pldmtool to separate package
Some system vendors may want to either:
    - only install pldmtool
    - prohibit installation of pldmtool

Split the pldm recipe so that it also generates a pldmtool sub-package,
which by default is still installed (via RRECOMMENDS) but can be
separately installed or explicitly prohibited.

Tested: Manually checked packages-split on pldm to confirm pldmtool
is placed in a separate package after build.  Built Yosemite4 and
confirmed pldmtool is still present.

Signed-off-by: Patrick Williams <patrick@stwcx.xyz>
Change-Id: I01fa9bd58de7bd1b9f65584ab61427ea74fe3bfb