meta-facebook: yosemite4: Add pldmd exit handler
Add a pldmd-exit-handler script and systemd service to set the virtual
sensors SENTINEL_DOME_SLOT_*_SENSOR_STATUS and
WAILUA_FALLS_SLOT_*_SENSOR_STATUS to 0 when pldmd exits or fails.

This is necessary because, by default, the virtual sensors do not
receive any signal or notification when the pldmd service exits or
fails.
As a result, they do not automatically re-read D-Bus sensor information
or update their values.

This ensures that when PLDM sensors are missing from D-Bus,
phosphor-pid-control enters fail safe mode and drives the fans to 100%,
helping to protect the system from thermal issues.

Motivation:
When phosphor-pid-control builds the fan table, any input sensor
missing from D-Bus will not be included in the fan table for
protection.
This means if the motherboard FRU is actually present in the system,
but pldmd crashes or for any reason the intended PLDM sensors are
missing from D-Bus, those sensors will not be protected.

Although phosphor-pid-control currently supports an option to enter
fail safe mode and drive the fans to 100% when input sensors are
missing from D-Bus, this function is disabled if unavailableAsFailed
is set to false.
See: https://gerrit.openbmc.org/c/openbmc/phosphor-pid-control/+/76062

However, in our system, we still need to set unavailableAsFailed to
false for the SENTINEL_DOME and WAILUA_FALLS sensors to ensure that the
system does not enter fail safe mode before POST completes.
Therefore, we still need to use the current virtual sensor approach to
achieve the desired protection.

Test:
On motherboard systems with either T1 or T2 configuration:
- Stop pldmd and verify the fan duty goes to 100%.
- Start pldmd and verify the fan duty returns to normal mode.

Related to:
https://gerrit.openbmc.org/c/openbmc/entity-manager/+/80270
https://gerrit.openbmc.org/c/openbmc/openbmc/+/80269

Change-Id: Ifd3d1a4ebdaa483abffe72b5801d036c33a89eba
Signed-off-by: Eric Yang <eric.yang.wiwynn@gmail.com>