meta-facebook: minerva: modify the fan monitor service sequence
We found that a false fan abnormality event log may occasionally appear
after BMC reboot. The root cause is that FCB_1 is not yet present when
the fan monitor service starts. This happens because the fan monitor
service starts before phosphor-gpio-presence, resulting in a failed
presence check for the FCB.

To resolve this, we changed the service startup sequence and added a
pre-execution sleep delay to give phosphor-gpio-presence enough time
to initialize all presence GPIOs.

Change-Id: If85ddf2f6c22acc9a90330145f68428da9061d56
Signed-off-by: Yang Chen <yang.chen@quantatw.com>