meta-facebook: yosemite4: Align fan IC watchdog
Aligned the watchdog timeout durations of different fan ICs
and extended them to 30 seconds:
- max31790:
10s -> 30s
bit[2:1]: 10 -> 11
- nct73633:
15s -> 30s
bit[3:2]: 00 -> 11

Motivation:
Currently, the watchdog timeout durations of different fan IC sources
are inconsistent.
If the watchdog timeout is too short, the fan IC may set the fan PWM
to 100% before the fan sensor service starts and initializes.
Although the Target property is later initialized to 100% after the
service starts, the actual fan PWM reverts to the configured value
(70% set by yosemite4-early-init).

This causes the fan sensor to report an unexpected Target property
(actual speed is 70%, but Target is 100%), which makes the
phosphor-pid-control logic fail when controlling the fan sensor to
reach 100% PWM at startup.

Change-Id: I6432a2964e5a85ce69763ef3971a2a8d2a973776
Signed-off-by: Eric Yang <eric.yang.wiwynn@gmail.com>