meta-facebook: ventura: add LED for FAN module location
Add support for identify LED control on the fan module to help locate
individual modules based on their presence and identify state.

When the fan module is present:
  - Asserted: LED blinks amber
  - Deasserted: LED is solid blue

When the fan module is not present:
  - Asserted: LED blinks amber
  - Deasserted: LED is solid amber

Tested:
Used busctl to toggle the Asserted property:
  busctl set-property xyz.openbmc_project.LED.GroupManager \
    /xyz/openbmc_project/led/groups/fcb0fan0_identify \
    xyz.openbmc_project.Led.Group Asserted b <true|false>
Confirmed LED behavior under both presence and absence conditions.

Change-Id: Ia87eedcdbf8c03ab2ea9fffd8dc0aab7f6a469d1
Signed-off-by: Fred Zhang <fred.zhang@quantatw.com>