meta-ibm: add logging-test to flash images
Our test team utilizes this for some automated test cases. Currently you
have to pass the debug tarball to the test suite so it can extract the
tool and scp it over to the BMC. Moving this test application into the
flash images allows us to simplify this process and stop building the
debug tarball with every build.

It's a small tool with limited flash use impacts.

Change-Id: Ieb4617003c2a4c6180694bfdcf38a49b6a657af6
Signed-off-by: Andrew Geissler <geissonator@yahoo.com>