meta-facebook: catalina: Add delay to debounce false RUN_POWER_PG
We observed that the BMC occasionally logs Host0 state is off twice,
even when there is no corresponding Host0 state is on log.

This issue is caused by a false RUN_POWER_PG assert and deassert (around
1 secons) while an unexpected HMC reset.

To address this, introduce a 2-second delay to debounce the RUN_POWER_PG
signal to avoid incorrect host state transitions logs.

Change-Id: I58af18e90d0b7b0bbecfa60c9cde7205d2d7d3df
Signed-off-by: Potin Lai <potin.lai@quantatw.com>