meta-facebook: ventura: persistent RS485 port mapping
Since the assignment of ttyUSB devices is not persistent to specific
physical ports, symbolic links are used to bind each port for easier
application access.
The symbolic link names are based on the labels printed on the tray,
with the prefix tty added for identification purposes.

The defined symbolic link names are as follows.
Front panel:
  ttyRS485-1, ttyRS485-2, ttyRS485-3

Rear panel: (The fab2 doesn't support ttyRS485-4 to ttyRS485-7)
  ttyRS485-4, ttyRS485-5, ttyRS485-6, ttyRS485-7,
  ttyMETER-0, ttyMETER-1, ttyMETER-2, ttyMETER-3

Tested:
The mapping between symbolic links and physical ports has been verified.

Change-Id: I7a40f7ef9e41dc3a0ac43bada54b55c1527948f2
Signed-off-by: P.K. Lee <p.k.lee@quantatw.com>