meta-google: gbmc-bridge: hybrid oob ip failover
This creates a special RA route oob-ip/124 ip so that tray can detect
whether the out-of-band IP is healthy or not. If the route exists then
it will prefer the oob as the src ip if metrics are same.

Tested: /124 route is added correctly. Tray can adjust the src IP based
on the route sent by the host. Tested both ncsi primary failover and
front primary failover for host and tray.

Change-Id: I59f93f5a69d90a48768f4b3efefdb5deead8a684
Signed-off-by: Yuxiao Zhang <yuxiaozhang@google.com>